pyproject.toml
[tool.poetry]
name = "template-drf"
version = "0.1.0"
description = ""
authors = ["陈俊武 <chenjunwu@dslyy.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"
drf-yasg = "^1.21.10"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
